% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActions.R
\name{getActions}
\alias{getActions}
\title{Solve a mathematical optimization problem using specified solver.}
\usage{
getActions(sol)
}
\arguments{
\item{a}{A list containing information about the optimization problem, including coefficients, constraints, and other parameters.}

\item{solver}{A character specifying the solver to be used ("gurobi" or "cbc").}

\item{cores}{An integer specifying the number of processor cores to be used.}

\item{verbose}{A logical value indicating whether to print solver output.}

\item{gap_limit}{A numeric scalar specifying the relative MIP (Mixed-Integer Programming) optimality gap as a stopping condition.}

\item{time_limit}{An integer specifying the time limit for the solver in seconds.}
}
\value{
A list containing the solution to the optimization problem, including variable values and solver statistics.
}
\description{
This function takes an optimization problem represented by a list of matrices and vectors and solves it using either the CBC (Coin-OR branch-and-cut) or Gurobi solver.
}
\examples{
solve(a, solver = "gurobi", cores = 2, verbose = TRUE, gap_limit = 0.0, time_limit = 100)
}
