% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{instance}
\alias{instance}
\title{Create an instance for simulation or from a file.}
\usage{
instance(
  type = NULL,
  file = "",
  n = 10L,
  species = 4,
  threats = 1,
  expansionType = NULL,
  threats_lloc = NULL,
  dlong = NULL,
  seed = 49
)
}
\arguments{
\item{type}{A character specifying the type of instance. Use "simulation" for simulation data or "file" for reading from a file.}

\item{file}{If type is "file", the path to the file containing the instance data.}

\item{n}{An integer specifying the number of units.}

\item{species}{An integer specifying the number of species.}

\item{threats}{An integer specifying the number of threats.}

\item{expansionType}{A matrix specifying the expansion type.}

\item{threats_lloc}{A numeric vector specifying threat locations.}

\item{dlong}{A matrix specifying distances.}

\item{seed}{An integer specifying the seed for reproducibility.}
}
\value{
A list representing the instance with various components.
}
\description{
This function generates an instance for simulation or reads from a file to create an object with various components.
}
\examples{
instance(type = "simulation", n = 10, species = 4, threats = 1)
instance(type = "file", file = "path/to/instance_data.txt")
x[between(x, -1, 1)]
}
