% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Create a dynamic model based on input parameters.}
\usage{
model(
  input = NULL,
  objective = "min_prop",
  levels = 3L,
  periods = 10L,
  budget = 0
)
}
\arguments{
\item{input}{A list containing the necessary components for creating the dynamic model. Components include Unidades, Especies, Amenazas, Jk, Ij, Ik, ExpansionType, Dlong, Adyacency, Dradial, and Ck.}

\item{objective}{A string specifying the type of objective to achieve in the mathematical model. There are two available: "min_prop" and "max_relative_benefit".}

\item{levels}{An integer specifying the number of levels in the model.}

\item{periods}{An integer specifying the number of periods for the dynamic model.}

\item{budget}{A numeric scalar specifying the budget for the model.}
}
\value{
A dynamic model based on the input parameters.
}
\description{
This function takes input parameters and uses them to create a dynamic model. The model is designed for dynamic actions, and the input should be a list containing various components.
}
\examples{
model(input_data, levels = 3, periods = 10, budget = 100)
}
